- isolates walletConnectorsComponent from the header file  - Removes other connectors since WalletConnect is working with all the wallet with want.  - Removes injected connector hook for eth since it is not used anymore.
