+++ src/components/layout/Header.js
@@ -1,4 +1,4 @@
-import React, {useState, useRef, useEffect} from "react";
+import React, {useRef, useState} from "react";
 import "./Header.css";
 import SearchRoundedIcon from "@material-ui/icons/SearchRounded";
 import "./SearchResults.css";
@@ -6,28 +6,9 @@ import {useHistory, useLocation} from "react-router-dom";
 import {useStateValue} from "../../StateProvider";
 import {motion} from "framer-motion";
 import Modal from 'react-modal';
-import {ethers} from 'ethers'
-import {
-    Web3ReactProvider,
-    useWeb3React,
-    UnsupportedChainIdError
-} from "@web3-react/core";
-import {BscConnector} from '@binance-chain/bsc-connector';
-import {
-    InjectedConnector,
-    NoEthereumProviderError,
-    UserRejectedRequestError as UserRejectedRequestErrorInjected
-} from "@web3-react/injected-connector";
-import {
-    URI_AVAILABLE,
-    UserRejectedRequestError as UserRejectedRequestErrorWalletConnect
-} from "@web3-react/walletconnect-connector";
-import {Web3Provider} from "@ethersproject/providers";
-import {formatEther} from "@ethersproject/units";
-
-import {injected, walletconnect, walletlink} from "../../connectors";
-import {useEagerConnect, useInactiveListener} from "../../hooks";
-import {Spinner} from "../misc/Spinner";
+import {Web3ReactProvider} from "@web3-react/core";
+import {ethers} from "ethers";
+import {WalletConnectorsComponent} from "../wallet/walletConnectorsComponent"
 
 function Header() {
     const location = useLocation();
@@ -64,331 +45,12 @@ function Header() {
         setIsOpen(false);
     }
 
-    const connectorsByName = {
-        Metamask: injected,
-        WalletConnect: walletconnect,
-        WalletLink: walletlink
-    };
-
-    function getErrorMessage(error) {
-        if (error instanceof NoEthereumProviderError) {
-            return "No Ethereum browser extension detected, install MetaMask on desktop or visit from a dApp browser on mobile.";
-        } else if (error instanceof UnsupportedChainIdError) {
-            console.log(error instanceof UnsupportedChainIdError);
-            return "You're connected to an unsupported network.";
-        } else if (
-            error instanceof UserRejectedRequestErrorInjected ||
-            error instanceof UserRejectedRequestErrorWalletConnect
-        ) {
-            return "Please authorize this website to access your Ethereum account.";
-        } else {
-            console.error(error);
-            return "An unknown error occurred. Check the console for more details.";
-        }
-    }
-
     function getLibrary(provider) {
         const library = new ethers.providers.Web3Provider(provider);
         library.pollingInterval = 12000;
         return library;
     }
 
-    function BalanceComponent() {
-        const {account, library, chainId} = useWeb3React()
-        const [balance, setBalance] = React.useState();
-        React.useEffect(() => {
-            console.log('Balance component effect')
-            if (!!library && !!account) {
-                let stale = false;
-
-                library
-                    .getBalance(account)
-                    .then(balance => {
-                        if (!stale) {
-                            setBalance(balance);
-                        }
-                    })
-                    .catch(() => {
-                        if (!stale) {
-                            setBalance(null);
-                        }
-                    });
-
-                return () => {
-                    stale = true;
-                    setBalance(undefined);
-                };
-            }
-        }, [library, account, chainId])
-
-
-        function formatBalance(balance) {
-            if (balance === null) {
-                return 'Error'
-            } else if (balance) {
-                return formatEther(balance);
-            } else {
-                return '';
-            }
-        }
-
-        return (
-            <>
-                <h3 style={{
-                    display: "grid",
-                    gridGap: "1rem",
-                    gridTemplateColumns: "1fr min-content 1fr",
-                    maxWidth: "20rem",
-                    lineHeight: "2rem",
-                    margin: "auto"
-                }}>
-                    <span>Balance</span>
-                    <span>{formatBalance(balance)}</span>
-                </h3>
-            </>
-        )
-    }
-
-
-    function AccountComponent() {
-        const {account} = useWeb3React()
-
-        function formatAccount(account) {
-            if (account === undefined)
-                return '...';
-            else if (account !== null)
-                return account.substring(0, 6) + '...' + account.substring(account.length - 4);
-            else
-                return 'None'
-        }
-
-        return (
-            <>
-                <h3 style={{
-                    display: "grid",
-                    gridGap: "1rem",
-                    gridTemplateColumns: "1fr min-content 1fr",
-                    maxWidth: "20rem",
-                    lineHeight: "2rem",
-                    margin: "auto"
-                }}>
-                    <span>Account</span>
-                    <span>
-                        {formatAccount(account)}
-                    </span>
-                </h3>
-            </>
-        )
-    }
-
-    function MyComponent() {
-        const context = useWeb3React();
-        const {
-            connector,
-            library,
-            chainId,
-            activate,
-            deactivate,
-            active,
-            error
-        } = context;
-
-        // handle logic to recognize the connector currently being activated
-        const [activatingConnector, setActivatingConnector] = React.useState();
-        React.useEffect(() => {
-            console.log('running')
-            if (activatingConnector && activatingConnector === connector) {
-                setActivatingConnector(undefined);
-            }
-        }, [activatingConnector, connector]);
-
-        // handle logic to eagerly connect to the injected ethereum provider, if it exists and has granted access already
-        const triedEager = useEagerConnect();
-
-        // handle logic to connect in reaction to certain events on the injected ethereum provider, if it exists
-        useInactiveListener(!triedEager || !!activatingConnector);
-
-        // set up block listener
-        const [blockNumber, setBlockNumber] = React.useState();
-        React.useEffect(() => {
-            console.log('running')
-            if (library) {
-                let stale = false;
-
-                console.log('fetching block number!!')
-                library
-                    .getBlockNumber()
-                    .then(blockNumber => {
-                        if (!stale) {
-                            setBlockNumber(blockNumber);
-                        }
-                    })
-                    .catch(() => {
-                        if (!stale) {
-                            setBlockNumber(null);
-                        }
-                    });
-
-                const updateBlockNumber = blockNumber => {
-                    setBlockNumber(blockNumber);
-                };
-                library.on("block", updateBlockNumber);
-
-                return () => {
-                    library.removeListener("block", updateBlockNumber);
-                    stale = true;
-                    setBlockNumber(undefined);
-                };
-            }
-        }, [library, chainId]);
-
-
-        // log the walletconnect URI
-        React.useEffect(() => {
-            console.log('WalletConnect effect hook running')
-            const logURI = uri => {
-                console.log("WalletConnect URI", uri);
-            };
-            walletconnect.on(URI_AVAILABLE, logURI);
-
-            return () => {
-                walletconnect.off(URI_AVAILABLE, logURI);
-            };
-        }, []);
-
-
-        return (
-            <div style={{padding: "1rem"}}>
-                <h1 style={{margin: "0", textAlign: "right"}}>
-                    {active ? "🟢" : error ? "🔴" : "🟠"}
-                </h1>
-                <AccountComponent/>
-                <BalanceComponent/>
-
-                <div
-                    style={{
-                        display: "grid",
-                        gridGap: "1rem",
-                        gridTemplateColumns: "1fr 1fr",
-                        maxWidth: "20rem",
-                        margin: "auto"
-                    }}
-                >
-                    {Object.keys(connectorsByName).map(name => {
-                        const currentConnector = connectorsByName[name];
-                        const activating = currentConnector === activatingConnector;
-                        const connected = currentConnector === connector;
-                        const disabled =
-                            !triedEager || !!activatingConnector || connected || !!error;
-
-                        return (
-                            <button
-                                style={{
-                                    height: "3rem",
-                                    borderRadius: "1rem",
-                                    borderColor: activating
-                                        ? "orange"
-                                        : connected
-                                            ? "green"
-                                            : "unset",
-                                    cursor: disabled ? "unset" : "pointer",
-                                    position: "relative"
-                                }}
-                                disabled={disabled}
-                                key={name}
-                                onClick={() => {
-                                    setActivatingConnector(currentConnector);
-                                    activate(connectorsByName[name]);
-                                }}
-                            >
-                                <div
-                                    style={{
-                                        position: "absolute",
-                                        top: "0",
-                                        left: "0",
-                                        height: "100%",
-                                        display: "flex",
-                                        alignItems: "center",
-                                        color: "black",
-                                        margin: "0 0 0 1rem"
-                                    }}
-                                >
-                                    {activating && (
-                                        <Spinner
-                                            color={"blue"}
-                                            style={{height: "25%", marginLeft: "-1rem"}}
-                                        />
-                                    )}
-                                    {connected && (
-                                        <span role="img" aria-label="check">
-                      ✅
-                    </span>
-                                    )}
-                                </div>
-                                {name}
-                            </button>
-                        );
-                    })}
-                </div>
-                <div
-                    style={{
-                        display: "flex",
-                        flexDirection: "column",
-                        alignItems: "center"
-                    }}
-                >
-                    {(active || error) && (
-                        <button
-                            style={{
-                                height: "3rem",
-                                marginTop: "2rem",
-                                borderRadius: "1rem",
-                                borderColor: "red",
-                                cursor: "pointer"
-                            }}
-                            onClick={() => {
-                                deactivate();
-                            }}
-                        >
-                            Deactivate
-                        </button>
-                    )}
-
-                    {!!error && (
-                        <h4 style={{marginTop: "1rem", marginBottom: "0"}}>
-                            {getErrorMessage(error)}
-                        </h4>
-                    )}
-                </div>
-
-                <div
-                    style={{
-                        display: "grid",
-                        gridGap: "1rem",
-                        gridTemplateColumns: "fit-content",
-                        maxWidth: "20rem",
-                        margin: "auto"
-                    }}
-                >
-                    {connector === walletconnect && (
-                        <button
-                            style={{
-                                height: "3rem",
-                                borderRadius: "1rem",
-                                cursor: "pointer"
-                            }}
-                            onClick={() => {
-                                connector.close();
-                            }}
-                        >
-                            Kill WalletConnect Session
-                        </button>
-                    )}
-
-                </div>
-            </div>
-        );
-    }
 
     const handleChange = (e) => {
         if (queryRef.current.value.length > 0) {
@@ -399,10 +61,6 @@ function Header() {
             setResultsOpen(false);
         }
     };
-
-    const styleButton = {
-        height: "2rem"
-    }
 
     return (
         <div className="header">
@@ -415,7 +73,7 @@ function Header() {
                 contentLabel="Example Modal"
             >
                 <Web3ReactProvider getLibrary={getLibrary}>
-                    <MyComponent/>
+                    <WalletConnectorsComponent/>
                 </Web3ReactProvider>
             </Modal>
             <div className="header__nav">
