--- src/connectors.js
@@ -1,25 +1,0 @@
-import {InjectedConnector} from "@web3-react/injected-connector";
-import {WalletConnectConnector} from "@web3-react/walletconnect-connector";
-import {WalletLinkConnector} from "@web3-react/walletlink-connector";
-
-const POLLING_INTERVAL = 12000;
-const RPC_URLS = {
-  1: "https://mainnet.infura.io/v3/84842078b09946638c03157f83405213",
-  4: "https://rinkeby.infura.io/v3/84842078b09946638c03157f83405213"
-};
-
-export const injected = new InjectedConnector({
-  supportedChainIds: [1, 3, 4, 5, 42, 56, 0x38]
-});
-
-export const walletconnect = new WalletConnectConnector({
-  rpc: { 1: RPC_URLS[1] },
-  bridge: "wss://relay.walletconnect.org",
-  qrcode: true,
-  pollingInterval: POLLING_INTERVAL
-});
-
-export const walletlink = new WalletLinkConnector({
-  url: RPC_URLS[1],
-  appName: "web3-react example"
-});
