--- src/utils/calls/index.js
@@ -1,3 +1,0 @@
-export * from './estimateGas'
-export * from './farms'
-export * from './pools'
