--- src/utils/providers.js
@@ -1,8 +1,0 @@
-import {ethers} from 'ethers'
-import getRpcUrl from 'utils/getRpcUrl'
-
-const RPC_URL = getRpcUrl()
-
-export const simpleRpcProvider = new ethers.providers.JsonRpcProvider(RPC_URL)
-
-export default null
